Add a CMake SDistTemplate
