+++ meta_dds/setup.py
@@ -11,7 +11,7 @@ from pathlib import Path
 from typing import List
 
 from meta_dds import cli
-from meta_dds.sdist import SDistTemplate, SDist, DirectoryPureSDist
+from meta_dds.sdist import SDistTemplate, ToolchainSpecificSDist, PureSDist
 from meta_dds.toolchain import DDSToolchain, get_dds_toolchain
 
 
@@ -25,7 +25,7 @@ class BuildSetup:
     def _meta_projects_dir(self) -> Path:
         return self.build_dir / '_meta_projects'
 
-    def _setup_meta_projects(self) -> List[SDist]:
+    def _setup_meta_projects(self) -> List[ToolchainSpecificSDist]:
         self._meta_projects_dir.mkdir(parents=True)
         return [sdist.instantiate(self.toolchain, self._meta_projects_dir / sdist.name)
                 for sdist in self.sdists]
@@ -34,7 +34,7 @@ class BuildSetup:
         self._setup_meta_projects()
 
 def build_setup_main(args: argparse.Namespace):
-    sdist = DirectoryPureSDist(args.project, [args.project / 'include'], [args.project / 'src'], [args.project / 'test'])
+    sdist = PureSDist(name='name', project_root=args.project)
     setup = BuildSetup(build_dir = args.output, sdists=(sdist,), toolchain=get_dds_toolchain(args.toolchain))
     setup.setup()
